Introduce 'operf --pid' -based profiling hook
This profiling hook is compatible with oprofile in its output, but allows
profiling only an interesting part of a single process instead of the whole
system. The monitor-pre-hook thus expects a third parameter with the PID.
Operf is started using a new script operf_start.sh started on background,
which starts operf, writes the operf's own PID in an operf.pid file and waits
for operf to finish. The monitor-post-hook reads the operf.pid to send
SIGINT to operf and waits for the removal of operf.pid file, which is done
after the startup script finishes waiting for operf to finish.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.de>