R support: enable R-based graphs
In compare-kernels.sh, reuse the --R parameter (introduced for comparisons)
also for plots. Also manage R_TMPDIR on the top level to allow reusing of
preprocessed results in multiple plots.

In graph-mmtests.sh, introduce --R parameter and also process -b parameter
so that it can be passed to R plotting. Also introduce --plottype parameter
to allow overriding extracted plottype.

The --R parameter is used to defer plotting to plot-R, with a fallback to the
classical plot script, for plot types not yet implemented in R. Currently
supported in R are boxplot and run-sequence plots.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.de>