This plugs the unixbench (http://code.google.com/p/byte-unixbench/) collection into mmtests. I've only chosen the system related workloads, specifically: dhry2reg, pipe, spawn, execl and syscall. Full description of each workload can be found in the project's site. All of them, however, involve blindly pounding at a high level various subsystems for a certain amount of time -- with parallel processes. This will obviously stress kernel internals, thus can be used for both small and large optimizations in the respective paths. At a higher level, results need to be taken with a grain of salt as behavior is pretty pathological and not real-world.
Note that the index number is pretty much ignored on this patch, and we
report only throughput (loops per second). Since:
(1) we don't run all benchmarks we can only generate partial indexes,
(2) the index strictly depends on the overall throughput, so we do not
miss out on any data. If lps increases, so does the index, and vice versa.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@suse.de>