This uses a simplified version of the "dedup" workload found in the PARSEC suite. It is the same core, just adapted to be easy to use and more unixy. For an overview of the actual workload, please consult the README in ezdedup, or even better, the PHD thesis that describes it in full.
In a nutshell: this is a deduplication (compression) workload,
which uses a 5 stage pipeline pattern for parallelism, each
stage using a producer consumer pattern between them. This
is interesting to us in two ways: IO since stages 1 and 5
deal with the input and output file images, respectively.
For the middle stages, the workload becomes in memory, pushing
heavily on the mmap sem as everything involves pthreads sharing
the address space. Ultimately, dedup is bounded by the output
I/O stage. The higher overall metric of interest here is wall
time -- the rest will depend on the input file, so compression
rates or ratios do not change. Other things to look at are obviously
page faults and the usual distribution of time.

Additionally, deduplication techniques have become a mainstream
method to compress storage footprints for new gen backup storage
systems. So having it in mmtests is quite useful from both a
performance pov and real-world impact.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@suse.de>