mmtests: Add data mining workload
Per the freqmine README:

""
Frequent Itemsets Mining (FIM) is the basis of Association Rule
Mining (ARM). Association Rule Mining is the process of analyzing
a set of transactions to extract association rules. ARM is a very
common used and well-studied data mining problem. The mining is
applicable to any sequential and time series data via discretization.
Example domains are protein sequences, market data, web logs, text,
music, stock market, etc.

To mine ARMs is converted to mine the frequent itemsets Lk, which
contains the frequent itemsets of length k. Many FIMI (FIM
Implementation) algorithms have been proposed in the literature,
including FP-growth and Apriori based approaches. Researches showed
that the FP-growth can get much faster than some old algorithms like
the Apriori based approaches except in some cases the FP-tree can be
too large to be stored in memory when the database size is so large
or the database is too sparse.
""

Since the freqmine workload uses the FP-growth based method, its use
of memory is unbounded and can incur in large working sets (as indicated
above). Use of memory will grow with the amount of data which is being
processed.

Some comments about the workload and integration for mmtests:

o Plug into mmtests is composed of three workloads: small, medium
  and large. The runtimes of the first two are quite marginal, the
  large will, however, take a substantial amount of time (200-20 secs)
  As such, the workload itself is highly parallelizable with increasing
  core counts -- although hyperthreading unfriendly. False sharing and
  cacheline invalidations looks to be an issue as large amounts of the
  ever changing tree will be in cache.

o Uses openMP for parallelism -- which ends up sharing the address space.
  Threads are scheduled dynamically, which means it uses an openMP
  internal workqueue to assign blocks of work to each thread using a producer
  consumer-like pattern. All this implies that there is extra overhead
  involved in each thread.

o Uses a custom version of the benchmark, isolated from parsec.

o Input files are provided in the benchmark installation.
  More details about input parameters are provided in comments
  in the `freqmine-bench' script.

o An FP-tree is a compressed form of the transaction database that
  can be traversed in multiple ways. This makes the program behavior
  highly dependent on the exact properties of the input data.

o Metric of interest is runtime, which is composed of (i) data preparation
  cost and (ii) FPgrowth cost (which does the actual mining). Note that
  the time for (i) is marginal compared to (ii). Experiments confirm this,
  thus we neglect it and report everything as one.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@suse.de>