netperf: Increase the UDP socket buffer
Previously we set the size of the UDP socket buffer to 32KB.
It seems not enough for 10Gbps ethernet cards and affected the result
seriously. It caused a big gap between the transmit rate and the
receive rate and the numbers varied from time to time.

This commit increases the size to 128KB. Per my experiment, the
difference between the transmit rate and receive rate is less than
0.5%, and the result becomes stable and reliable.

Signed-off-by: Gary Ching-Pang Lin <glin@suse.com>
Signed-off-by: Mel Gorman <mgorman@suse.de>