Dashboard: Determine the goodness of result by the n-sigma difference
Before this commit, all results in dashboard are compared by simple
ratio and then we check if the geometric mean is better or worse than
before.

From now on, when it is possible (there is some data to count stddev
from), we now check the difference between new and base to the stddev of
the base mean.

Typically, the tests are running more times, we get several numbers
which are then counted into some kind of mean and we can also have
stddev of these values.

The premise here is to check the new result inside some interval around
the base value, typically inside <base-sigma*k, base+sigma*k> for some k
(chosen k=1 for now). This means that the difference has little
statistical significance.

We still show the ratio as it is more readable for humans but we color
them by stddev difference to see if the eg. 1% change is bad or ok.

Signed-off-by: Jan Moskyto Matejka <mq@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.de>