mmtests: add will-it-scale
The will-it-scale (wis) collection is a nice microbenchark suite,
which stresses multiple areas of interest to kernel developers.
There are currently 51 programs, in both threads and processes
flavors (called model in the scripts). Add support for it to mmtests,
grouping as follows:

wis-futex
	futex1
	futex2
	futex3
	futex4
wis-seek
	lseek1
	lseek2
	readseek1
	readseek2
	readseek3
	writeseek1
	writeseek2
	writeseek3
wis-pread
	pread1
	pread2
	pread3
wis-pwrite
	pwrite1
	pwrite2
	pwrite3
wis-malloc
	brk1
	malloc1
	malloc2
wis-pf
	page_fault1
	page_fault2
	page_fault3
wis-unlink
	unlink1
	unlink2
wis-mmap
	mmap1
	mmap2
wis-poll
	poll1
	poll2
wis-filelock
	lock1
wis-open
	open1
	open2
	open3
wis-read
	read1
	read2
	read3
	read4
wis-getppid
	getppid1
wis-sched
	sched_yield
wis-fallocate
	fallocate1
wis-pthreadmutex
	pthread_mutex1
	pthread_mutex2
wis-signal
	signal1
wis-pipe
	context_switch1
	pipe1
wis-eventfd
	eventfd1
wis-posixsems
	posix-semaphore1

The benefits of such a logical order are (i) that we can run
isolated benchmarks of interest -- ie: wis-futex can be added
to our current futex workloads; (ii) it avoids contaminating
the mmtests tree with endless wis-specific files.

The downside, of course is that we don't get monitor granularity
for each individual run (or model), only the whole group. In a
similar manner, this also affects the graphs for html reports,
which are currently disabled.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@suse.de>