dashboard: Dump top bad data
In order to compute the TOP10 at the beginning of dashboard, all the
dashboards must dump the data in some simple parseable format.

Adding three commandline arguments to dashboard-kernels-*. The static
version ignores these arguments (but must know about them to avoid
complaining), the smart version uses them accordingly.

This commit is to be followed with a matching commit in marvin repo.
This commit is to be the first as it shouldn't break anything when
called from old marvin. The marvin commit, on the other hand, relies on
this commit.

Signed-off-by: Jan Moskyto Matejka <mq@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.de>