Add pi_stress workload
This goes well in the futex related config file, such that we can
collect numbers for priority aware rtmutexes (PTHREAD_PRIO_INHERIT).

The workload will create 3 threads per inversion groups, with low,
medium and high priorities, which will attempt to take a pthread
mutex with PTHREAD_PRIO_INHERIT attributes and play games with
preempting the tasks that hold the mutex such that the PI is properly
stressed  for a determined amount of time (seconds).

Note that I've kept the default max inversion groups small, just to
the number of cpus in the system as overcommitting can cause deadlocks
and thus break our automation. Nonetheless, it is still a valuable
workload to have.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@suse.de>