mariadbbuild: allow users to pass startup options
It is quite normal for different parameters to be set for
different needs, DB tuning will usually depend on workload
and hardware characteristics. As such, instead of blindly
modifying my.cnf, allow the mariadb shellpack user to pass
startup options, in the form of a single string:

     "--option1=A,--option2=B,--option3=C"

Specific parameters to be passed to mysqld startup scripts
can be found:

https://dev.mysql.com/doc/refman/5.7/en/server-options.html

Another benefit of this is that it reduces work having to
parse the ini file every time -- this will only be necessary
for those parameters that cannot be used in this fashion.

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>
Signed-off-by: Mel Gorman <mgorman@suse.com>