shellpacks: add wikimedia web application
This adds the entire mediawiki application (the software
behind wikipedia, among others) in the form of a shellpack to be easily used
by future benchmarks (mainly wikibench). There are a number of caveats
regarding certain versions and methods of installation. Similarly, lots of
comments in the shellpack_src/ that can permit easier transitions.

In a nutshell:
   (1) Using the shellpack automatically installs apache and mariadb.
   (2) Download and configure multiple packages and the bzip2'ed xml
       dump from wikipedia.
   (3) Load the dump (with proper database innodb tunning).

For now the chosen version is based on 1.18 and will create a wikipedia
mirror reflecting mid 2008 (14 Gb uncompressed), with over 6.2 million
pages. Importing the database can take from 45 min up to 12 hrs, depending
on the hardware. Note that nosync option is used for flushing to disk
so expect shutdowns to be a little slower. Other than that this option
is by far the best for this kind of benchmark and setup.

Additionally, to have more data, avoid 404s and make better use of
related benchmarks, also add extras DB dumps for metadata, links and
images, among others. As the comments indicate, these can easily be
commented out if lack of disk space. The total installation size
requires over 120Gb.

Note that DB import phase could easily be considered a separate benchmark.

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>
Signed-off-by: Mel Gorman <mgorman@suse.com>