mmtests: add wikibench
This patch builds on top of mediawiki shellpack to make use
of the wikibench tool. This benchmark allows systems which
service the web and database servers to stressed at different
levels, indicating the points of where the resource exhaustion
Such scenarios can trigger missed deadlines or http 500 errors,
and therefore the request was not able to be serviced/timeout.
Software is entirely written in java.

The benchmark model is master/server, or controller/worker in
wikibench terms. The controller (single task) will feed trace
lines to one or more workers and schedule work. Each worker will
parse the url/timestamp from the read line and perform a GET or
POST request to the SUT (System Under Test -- which is where the
controller is running -- for now everything is localhost, but is
easily configurable) at the indicated time.

Workloads/Sampled Files
-----------------------

Given the above, most of the thought behind this benchmark is
generating or having, the actual traces of interest. This will
obviously determine the 'when and how' the SUT will be hit with
requests. Wikpedia provides previously sampled traces, mainly for
the months of 09 and 10 of 2007 -- there are other dates but links
are very fragile. These are real world traces and therefore are
very much real work and of interest.

One of the first steps to use wikibench is to prepare the vanilla
downloaded samples -- sorting, removing invalid and sampling to
reduce traffic. This last one can be done either via page or request
level sampling. I have setup a number of files in

http://linux-scalability.org/benchmarks/

that contain previously sampled and working traces for the month of
September of 2007. Unfortunately for now, one of the tools provided
by wikibench to perform page-level sampling is broken, and therefore
request level is only provided. The reduction permil (threshold) is
another configurable parameter, and I have created 3 files with each
size: small, medium, large. To aid debugging, another option that
overrides the default sizes allow passing the exact workload to use
for the traces.

Key Performance Metrics
-----------------------

At a high level:

(1) The main one is that the amount of successfully completed
requests. The closer this value is to 0, the benchmark ran
well under the current hardware. This is determined by verifying
the amount of 500 and the explicit missed deadlines from the bench
controller output.

(2) Mean response time of successfully completed requests.

(3) Request rate (requests/min).

There are plenty of comments along the benchmark pack and its
configuration.

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>
Signed-off-by: Mel Gorman <mgorman@suse.com>