libmicro-*: change all -I's to -B 10000
This change replaces -I <value> with -B 10000 in all invocations. A
sample size of 10000 has proven to be large enough for the faster
functions, like getpid(2), so it's good for all tests -- provided we
have enough time to run them (the problem of test duration will be
addressed in a separate patch).

The motivation is to circumvent a bug in libmicro. libmicro tries to
compute the clock frequency of a machine in a way that isn't
applicable to nowadays computers.

In order to explain the rationale of the change, here an illustration
of the semantics of the option flags -I, -B and -C and a recall of how
libmicro runs an experiment. From the libmicro help (accessible
calling any benchmark with -?):

    ./getpid -?
    usage: getpid
	   [... snip ...]
	   [-B batch-size (default 0)]
	   [-C minimum number of samples (default 0)]
	   [...]
	   [-I] nsecs per op (used to compute batch size)
	   [...]
    notes: measures getpid()

An experiment, or benchmark, consists of several "samples". A sample
consists of several calls to the function under test batched together.
The flag -B tells libmicro how many calls to perform in a sample or,
in other words, the size of a sample. An experiment is completed after
a fixed amount of time has passed, or a given number of samples has
been collected. The latter parameter is set via the -C option flag.

The value sample_size (also referred to as "batch size", set via -B)
should be, loosely speaking, inversely proportional to how fast a call
to the test function is: you want to batch many calls to a fast
function in the hope that after a sample has completed the clock will
have moved forward. Also, the higher the resolution of your timer is,
the smaller you can keep your samples: a few calls will suffice to
observe a time increment.

The flag -I offers the user an indirect way to specify the sample size;
the user is supposed to give libmicro a "hint", which I'll name
call_duration, of how long a single call to the function under test
will last. libmicro will then use it to compute the sample size with

    sample_size = time_resolution * 100 / call_duration

The libmicro bug lies in how time_resolution is computed: instead of
using clock_getres(2), it performs an increasingly larger number of
no-ops until it registers a time increment (see the function
get_nsecs_resolution() in libmicro.c at [0]). This naive way of
essentially measuring the clock frequency cannot work on modern
machines because of frequency scaling, and will result in a computed
sample_size which is often too small to be statistically significant.

My experiences report that time_resolution as computed by libmicro is
always either 1 nanosecond or 1000 nanosecond, without any apparent
correlation to the CPU type or nominal clock frequency. More over,
subsequent invocations of this function can give different results
*on the same machine*.

You can easily try running multiple times

    $ bin-x86_64/tattle -r

which outputs just that (tattle is a libmicro binary whose only
purpose is to give information on the library, such as the version of
the compiler used, and so forth).

time_resolution affects only another part of the code, where it's used
to signal potential quantization errors. If data is collected more
often than time_resolution * 100, the warning will be present in the
output; this to say that vital functions of the tests aren't affected
(and the "quantization errors" warning is, more often than not,
bogus).

[0] https://java.net/projects/libmicro/sources/hg-repo/content/libmicro.c?rev=14

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>