libmicro-*: make tests stop after 30 seconds.
This change makes tests stop after 30 seconds have passed, no matter
how many samples have been collected. Our tests needs to have
predictable duration, as we cannot afford to be stuck in a scenario
and delaying the whole campaign.  This is achieved by the option flags
-C 0 -D <duration> combined.

We'll now illustrate the semantics of option flags -C, -D and -B and
present the reason why simply setting -D <duration> is not sufficient
to achieve the desired result.

From libmicro help:

    ./getpid -?
    usage: getpid
	   [...]
	   [-B batch-size (default 0)]
	   [-C minimum number of samples (default 0)]
	   [-D duration in msecs (default 10s)]
	   [...]
    notes: measures getpid()

The flag -B sets the sample size, -C is for the number of samples
required, -D is a time duration. Running

    ./SOME-FUNC -D 30000 -B 10000

which is, setting the sample size to 10 thousands calls of SOME-FUNC
and the duration to 30 seconds, will not necessarily lead to a test
duration of 30 seconds: libmicro will run until it collects at least
100 samples, i.e. 10000*100 calls of SOME-FUNC. This is in contrast to
what advertised in the help screen, that says the default for -C is
zero. Here the stop condition (from libmicro source):

    if (elapsed_time > duration &&
        (!min_samples || min_samples < samples_collected)) {
        stop_running();
    }

min_samples can be provided by the user with -C, but is otherwise
initialised to 0. The way to enforce D as a "hard deadline" is to run

    ./SOME-FUNC -C 0 -D 30000 -B 10000

but at this point we are on our own with respect to how many samples
get collected.

It is observed that most of the tests collect enough 10k-sized samples
in 30 seconds. Those who don't will be moved to a different test suite
in a forthcoming commit, where the sample size will be smaller. We
prefer not to use ad-hoc settings for each invocation.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>