libmicro-regular: add test for exit()
All tests present in libmicro before this refactoring have been
reintroduced, apart from this one.

Before this refactoring our libmicro invocations contained the
following three lines:

    exit	exit_10		-B 10
    exit	exit_100	-B 100
    exit	exit_1000	-B 1000

which is, we ran the test 3 times with different sample sizes. This
can be traced back to the original libmicro repository, see the script
bench.sh at [0], but doesn't make much sense. If one has the time, the
larger the sample size the better.

[0] https://java.net/projects/libmicro/sources/hg-repo/content/bench.sh?rev=14

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>