libmicro-*: remove tests that use the /var directory
Several I/O tests were present in two variations: one writing/reading
on /tmp, and one writing/reading on /var. As this oddity appears in
the original libmicro repository [0], we imagine once upon the time
this choice was motivated. In our systems (openSUSE) the two
directories are on the same filesystem type (btrfs), so the tests are
duplicates.

[0] https://java.net/projects/libmicro/sources/hg-repo/content/bench.sh?rev=14

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>