dbench: Reduce maximum number of clients
Currently dbench uses 64 * NUMCPUS number of clients. This is a fine
stress test but it bumps necessary filesystem size which increases test
variance. Reduce number maximum number of clients to 8 * NUMCPUS which
should still provide good "overload" conditions.

Update dbench driver to make sure a config with number of clients >>
number of cpus is run even for this setting of maximum number of
clients.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>