dbench: Runtime scaling
Currently we run dbench for 15 minutes for each number of clients. That
is unnecessarily long for low numbers of clients (there 3 minutes are
plenty) and too short for high numbers of clients especially in 'sync'
configs. Scale dbench runtime based on number of clients. We linearly
scale the runtime as the number of clients grows. The baseline runtime
has to be configured, the factor of the scaling can either be configured
or it is chosen as the time it takes to run workload for one client.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>