dbench: Separate warmup and execution
Reported dbench throughput is somewhat a gamble. The original load file
consists of a preamble and then ~130 repetitions of a single workload
(with only tiny differences). Since the whole load file takes a long
time (hundreds to thousands of seconds) to execute for high numbers of
clients it is a gamble how close to the preamble the execution terminates
and that results in variation of throughput as the througput while executing
preamble is widely different from the throughput during the rest of the
loadfile. The throughput also varies considerably during the repeating load
so depending on in which phase the load the benchmark terminates, we get
a different throughput. This is mitigated when the workload executes
enough times but for large numbers of clients the time to execute the
workload enough times takes exceedingly long.

Switch dbench to execute preamble first to create the directory
hierarchy and then execute the basic workload in a loop. That removes
the variance from executing preamble. Also ask dbench to report times to
execute the full workload - that removes the variance from where in the
workload we terminate the benchmark.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>