dbench: Add parsers for dbench time output
So far we have used throughput numbers output by dbench to do statistics
for comparison. That does not really make a good sense as those numbers
are just cumulative throughput computed in the run so far. Also
throughput numbers are noisy as throughput is not uniform during the
load file execution. Move dbench4 reporting to use times to execute the
load file instead.

Since tbench4 uses the same reporting as dbench4, copy the original
extraction and comparison modules there for now. Later we may want to
switch tbench4 to time reporting as well.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@suse.com>