Add Chris Mason's schbench benchmark
schbench is based on the workloads that facebook run and allows us to
measure scheduler wakeup latency for non-realtime tasks.

The bin/lib/MMTests code is slightly larger than for other benchmarks
due to schbench providing its own quartile data.

A new config file is added in configs/ rather than adding schbench to
the existing scheduler config files so that existing tests do not need
to be deleted and re-run. The values for thread and messenger thread
count (SCHBENCH_THREADS and SCHBENCH_MESSAGE_THREADS, respectively)
are intended to mirror how facebook run their machines, i.e they are
not overloaded like crazy with more tasks than cpus.

One of the features of schbench that isn't wired up yet is "auto mode"
which grows the thread count automatically until the p99 latencies
become "horrible". This reason this isn't used is because it doesn't
easily fit into the existing mmtests reporting framework.

Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>