configs: Split scheduler configs and run lmbench separately
lmbench has shown unstable results in the past. Split it into its own
config so it can be evaluated separetly. If changes are requried, it's
relatively easy to rerun just lmbench. A different range of process sizes
are chosen to see if some of the instability was due to caching effects.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Acked-by: Matt Fleming <matt@codeblueprint.co.uk>