ipcscale: allow more workload options
We currently blindly rely on the benchmark to workout the
run options and determine system topology, which is fine,
except that at times a bit inflexible, such as when specifying
the number of threads running on a given core (ie when HT is of
interest). In addition, there are two different workloads that
the command offers when passed the -o option, both perform basic
ping-pong operations on a set for regular sysv and posix flavors.

Specifically, this patch:

(1) Adds MIN/MAX thread config options which determine the threads
    per core.
(2) Adds basic RUNTIME option.
(3) Splits the workload into three (i) waitforzero, (ii) sysvsempp
    and (iii) posixsempp.

There is additional info about the workload and options in the
config file.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>