ExtractPipetest: Round times to nearest half microsecond
Using the nanosecond resolution currently emitted by pipetest when
comparing results gives the impression we care about such miniscule
changes in runtime, but that's simply not true.

Given that we don't pin pipetest to CPUs out of the box, fluctuations
in runtime on the order of hundreds of nanoseconds are to be expected
in the face of migrations across CPUs.

It's unlikely that these small differences are caused by performance
regressions, so let's round the values to the nearest 500 nanoseconds
to smooth out the runtime variation and stop the mmtests reporting
framework constantly flagging them as potential issues.

Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>