Adrestia: Add new scheduler load balancer test suite
This patch wires up the wakeup-periodic and wakeup-single tests into
their own config files so as not to disturb existing archived results.

wakeup-single measures the time (90th percentile) to wakeup a single
thread. It basically measures the best case (minimum) time to perform
a wakeup.

wakeup-periodic takes service (spin) time and arrival (time between
wakeups) parameters, and so nicely approximates a "bursty" workload.
A range of arrival times and thread counts are used to monitor how it
affects wakeup latency.

A service time of 25ms was chosen on the assumption that it's long
enough to cause changes in CPU frequency and also short enough to
avoid excessively long test times.

More tests from the adrestia test suite will be added in the future.

Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>