apachebuild: Make httpd run as the user who starts it
For the purpose of some benchmarks (i.e. siege) httpd will serve documents that
are located somewhere under the mmtests/ directory. If this is under the $HOME
of some $USER we might incur in the situation where only $USER can read it,
because $HOME doesn't have the exec bit set, necessary for directory listing.

Before this patch, httpd is run under a specific user called "daemon", set in
httpd.conf with the User and Group directives. The result is

    # curl -I http://localhost/siege.html
    HTTP/1.1 403 Forbidden

This patch removes such directives alltogether with the effect of httpd not
switching to a new user but running under the same $USER that invoked
`apachectl -k start`.

The apache 2.4 documentation states that the default value for the User and
Group directive (which this change falls back to) are both #-1. It's not
entirely clear to me what happens if one calls setuid(-1) and setgid(-1), but
this config "works on my machine(TM)".

<https://httpd.apache.org/docs/2.4/mod/mod_unixd.html#User>
<https://httpd.apache.org/docs/2.4/mod/mod_unixd.html#Group>

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>