dashboard: round values to int before evaluating arith expression
Fix a bug by which a floating point number could be fed to a $((...))
bash expression. This resulted in a syntax error and consequent early
exit from a for loop responsible for generating report results.

We modify both the new and old reporting scripts
(format-for-display.sh and dashboard-kernels-smart.sh) despite the
latter being on its way to deprecation, as otherwise it would be
difficult to compare the output of the two during the validation
phase.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>