dashboard: remove obsolete scripts
To recap the history of these files: the original reporting script is
dashboard-kernels.sh, created on 10/2014 with 6b04877 "reporting: Add
support for printing ratios and an overall evaluation".

Around 04/2015 new features were implemented, but instead of modifying
dashboard-kernels.sh it was decided to rename it into
dashboard-kernels-static.sh (56d8ab4 "Rename dashboard-kernels.sh in
preparation for replacing it"), then add the desired changes to
dashboard-kernels.sh (92fad3a "Dashboard: Determine the goodness of
result by the n-sigma difference"), and finally to rename
dashboard-kernels.sh again into dashboard-kernels-smart.sh (b6ed90c
"dashboards: Preserve methods for generating dashboard but default to the
new version"). At the end of this operations, dashboard-kernels.sh was
made a symlink poiting to dashboard-kernels-smart.sh.

My personal understanding is that the file dashboard-kernels-static.sh
was meant to be removed after the new version proved its value, but the
removal never happened and the file has slowly drifted into
obsolescence.

As this commit series implements cleanups after an even further evolution
(where format-for-display.sh replaces dashboard-kernels-smart.sh), we
take the chance to remove those other scripts which are no longer
maintained.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>