json export: utility function to unbless objects
Get an unblessed reference to the content of the data structure
underlying an object. Will be used by the Compare.pm and Extract.pm
modules, for the purpose of their JSON serialization. JSON
serialization methods refuse to treat blessed objects.

We avoided the more natural name "unbless" as this is already taken by
a function in the popular Data::Structure::Util module. Such function
can't be used in this case as it acts in-place on the argument, so it
would require copying objects and we don't have enough memory for that.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>