json export: JSON serialization of Compare.pm and Extract.pm objects
The JSON serialization function "encode" from the Cpanel::JSON::XS
module requires blessed objects to implement their own TO_JSON method,
which is responsible for returning an unblessed reference for
Cpanel::JSON::XS->encode to work with.

The serialization is hidden behind the --json-export command line
flag, so that it can be enabled only if desired.

The output file name is passed to the serialization method; ideally
one could have used the value of $self->{_ModuleName} to determine the
name of the produced file, but many of the Compare/Extract subclasses
do not follow the convention of overriding their _ModuleName and
fixing that would have required hundreds of lines of diff generated
with a script.

The entry point of the serialization procedure is the Compare.pm
module; this might raise an eyebrow as a more natural choice would
have been directly serializing the Extract objects, which are
contained in the Compare objects. Specifically, a Compare object
contains one or more benchmark results in its
$self->{_ExtractModules} field, which is a list of Extract objects.
An Extract object merely parses a benchmark result.

The Compare object itself selects the representative data points in
its _ExtractModules field and applies statistical functions such as
quantiles computation or geometric/arithmetic mean.

By capturing the whole content of a Compare object one can have
already at hand those added statistics, which can allow to evaluate
their quality and identify possible errors or blind spots of the
current statistical routines.

What follows is the (pretty printed) content of a json export of the
data structure that underlies the Compare object for the benchmark
sqlite (configuration global-dhp__db-sqlite-insert-medium-ext4)
comparing kernels 4.4 vanilla and 4.9 vanilla. The names of the fields
are generally self-explanatory; when they aren't, the can be grepped
in the MMTests source code.

When analyzing a json export you'll find this dump of great help, as
it illustrates the hierarchy of fields (or schema, abusing the term)
showing where things are.

    {
      _CompareLength      => 6,
      _CompareOp          => "pdiff",
      _CompareTable       => [[0, -12.4253326234477]],
      _DataType           => 16,
      _ExtractModules     => [
			       {
				 _CompareLookup   => {},
				 _DataType        => 16,
				 _FieldFormat     => ["%-12s", "%12d", "%12.2f", "%12.2f", "%12d"],
				 _FieldHeaders    => ["Type", "Sample", "Ops"],
				 _FieldLength     => 12,
				 _ModuleName      => "SummariseVariabletime",
				 _Operations      => ["Trans"],
				 _PlotHeaders     => [],
				 _PlotLength      => undef,
				 _PrintHandler    => {},
				 _RatioPreferred  => "Higher",
				 _ResultData      => [
						       ["Trans", 0, 414.9738805045],
						       ["Trans", 1, 415.974214244913],
						       ["Trans", 2, 289.041796506834],
						       ["Trans", 3, 402.020769486782],
						       ... data here ...
						     ],
				 _SummariseColumn => 2,
				 _SummaryData     => [["Trans", 474.967726277831]],
				 _SummaryHeaders  => ["Time", "Ratio"],
				 _SummaryLength   => 16,
				 _TestName        => "4.4.0-vanilla",
			       },
			       {
				 _DataType        => 16,
				 _FieldFormat     => ["%-12s", "%12d", "%12.2f", "%12.2f", "%12d"],
				 _FieldHeaders    => ["Type", "Sample", "Ops"],
				 _FieldLength     => 12,
				 _ModuleName      => "SummariseVariabletime",
				 _Operations      => ["Trans"],
				 _PlotHeaders     => [],
				 _PlotLength      => undef,
				 _PrintHandler    => {},
				 _RatioPreferred  => "Higher",
				 _ResultData      => [
						       ["Trans", 0, 415.974214244913],
						       ["Trans", 1, 355.966476736471],
						       ["Trans", 2, 355.958415121219],
						       ["Trans", 3, 355.966137292676],
						       ... data here ...
						     ],
				 _SummariseColumn => 2,
				 _SummaryData     => [["Trans", 415.951406433784]],
				 _SummaryHeaders  => ["Time", "Ratio"],
				 _SummaryLength   => 16,
				 _TestName        => "4.9.0-vanilla",
			       },
			     ],
      _FieldHeaders       => [],
      _FieldLength        => 12,
      _GeometricMeanTable => [1, 0.875746673765523],
      _ModuleName         => "CompareSqlite",
      _PrintHandler       => {},
      _ResultData         => [],
      _ResultsRatioTable  => [[1, 0.875746673765523]],
      _ResultsTable       => [[474.967726277831, 415.951406433784]],
      _TitleTable         => [["Ratio", "Trans"]],
    }

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>